Fixed problems with equipment.isWearing() due to circular dependency
